feat(settings): Add settings view to protected routes
The 'settings' view has been added to the list of protected views in the App component. This ensures that users must be logged in to access the settings page. Additionally, a new LikesView component has been introduced, which allows users to view posts they have liked and posts from profiles they have liked. The UI for this view includes tabbed navigation for 'profiles' and 'posts'.