feat: Add story deletion functionality
Implements the ability to delete stories from the `StoryContext` and integrates it into the `StoryViewerModal` and `CreateStoryPostPage`.

The `FixedPageFooter` component has been updated to accept `children` as a prop, allowing for more flexible placement of additional elements within the footer. This change enables the addition of a delete button in the story editing view.